+++ src/main/java/com/codeborne/selenide/conditions/And.java
@@ -14,12 +14,26 @@ import static java.util.stream.Collectors.joining;
 @ParametersAreNonnullByDefault
 public class And extends Condition {
 
-  private final List<Condition> conditions;
+  private final List<? extends Condition> conditions;
   private Condition lastFailedCondition;
 
-  public And(String name, List<Condition> conditions) {
-    super(name, conditions.stream().allMatch(Condition::applyNull));
+  /**
+   * Ctor.
+   *
+   * @param name       condition name
+   * @param conditions conditions list
+   * @throws IllegalArgumentException if {@code conditions} is empty
+   */
+  public And(String name, List<? extends Condition> conditions) {
+    super(name, checkedConditionsListCtorArg(conditions).stream().allMatch(Condition::applyNull));
     this.conditions = conditions;
+  }
+
+  private static List<? extends Condition> checkedConditionsListCtorArg(List<? extends Condition> conditions) {
+    if (conditions.isEmpty()) {
+      throw new IllegalArgumentException("conditions list is empty");
+    }
+    return conditions;
   }
 
   @Nonnull
@@ -32,7 +46,6 @@ public class And extends Condition {
   @Override
   public boolean apply(Driver driver, WebElement element) {
     lastFailedCondition = null;
-
     for (Condition c : conditions) {
       if (!c.apply(driver, element)) {
         lastFailedCondition = c;
@@ -52,7 +65,6 @@ public class And extends Condition {
   @CheckReturnValue
   @Override
   public String toString() {
-    String conditionsToString = conditions.stream().map(Condition::toString).collect(joining(" and "));
-    return String.format("%s: %s", getName(), conditionsToString);
+    return getName() + ": " + conditions.stream().map(Condition::toString).collect(joining(" and "));
   }
 }